+++ src/Serilog.Sinks.Raygun/Sinks/Raygun/RaygunSink.cs
@@ -48,7 +48,6 @@ public class RaygunSink : ILogEventSink
     private readonly string _userInfoProperty;
     private readonly Action<OnBeforeSendArguments> _onBeforeSend;
     private readonly IEnumerable<Type> _wrapperExceptions;
-    //private readonly IEnumerable<string> _ignoredFormFieldNames;
 
     private readonly RaygunSettings _settings;
     private readonly IRaygunClientProvider _raygunClientProvider;
@@ -93,7 +92,6 @@ public class RaygunSink : ILogEventSink
         _userInfoProperty = userInfoProperty;
         _onBeforeSend = onBeforeSend;
         _wrapperExceptions = wrapperExceptions;
-        //_ignoredFormFieldNames = ignoredFormFieldNames;
 
         _settings = settings ?? new RaygunSettings
         {
@@ -121,12 +119,6 @@ public class RaygunSink : ILogEventSink
         {
             client.AddWrapperExceptions(_wrapperExceptions.ToArray());
         }
-
-        // TODO: Figure out how we can 'ignore' form field names on mobile platforms
-        // if (_ignoredFormFieldNames != null)
-        // {
-        //     client.IgnoreFormFieldNames(_ignoredFormFieldNames.ToArray());
-        // }
 
         client.CustomGroupingKey += OnCustomGroupingKey;
 
@@ -165,7 +157,12 @@ public class RaygunSink : ILogEventSink
         // Submit
         if (logEvent.Level == LogEventLevel.Fatal)
         {
+#if NETFRAMEWORK
             client.Send(exception, tags, properties);
+#else
+            // Fail is prob going to crash the app so let's try to send the log to raygun synchronously
+            client.SendAsync(exception, tags, properties).GetAwaiter().GetResult();
+#endif
         }
         else
         {