+++ CHANGE-LOG.md
@@ -1,5 +1,18 @@
 # Full Change Log for Serilog.Sinks.Raygun package
 
+
+### v8.0.0
+- Updated Raygun4Net dependency
+- Added `ApplicationBuilderExtensions` to allow `AddRaygun(...)` for use in Console/Services
+- Implemented new `RaygunClientSink` which is cut down but supports using whatever RaygunClient is registered via DI
+  - This treats RaygunClient as a singleton and uses it to send messages
+  - Adds support for using Raygun4Net.AspNetCore package to register RaygunClient in DI
+  - Adds support for `IRaygunUserProvider` to be registered in DI
+  - See: https://github.com/MindscapeHQ/serilog-sinks-raygun/pull/71
+- Notes: RaygunSink is left unchanged, this should not be a breaking change for anyone using the package as it is, 
+         but the new `RaygunClientSink` is available for those who wish to use it. The new sink is more flexible and allows 
+         for custom RaygunClient instances to be used.
+
 ### v7.5.0
 - Fixed issue in RaygunSink where it was not setting Request/Response messages if supplied by the code Enricher
 - Bumped Raygun.NetCore dependency to 8.2.0 to fix issue where custom RaygunClient with HttpClient could not be used